organized the code in the notes package by separating the implementations in their own source file.
