+++ cmd/app/notes/edit_note.go
@@ -10,6 +10,8 @@ import (
 	"notebook_app/cmd/app/notebook_db"
 
 	"notebook_app/cmd/app/user_session"
+
+	"notebook_app/cmd/app/form"
 )
 
 func EditNoteForm(db *gorm.DB) gin.HandlerFunc {
@@ -38,6 +40,32 @@ func EditNoteForm(db *gorm.DB) gin.HandlerFunc {
 			"note_description": note.Description,
 			"user":				user,
 		})
+
+	}
+	return gin.HandlerFunc(fn)
+}
+
+func EditNote(db *gorm.DB) gin.HandlerFunc {
+	fn := func(c *gin.Context) {
+		/* Edit Note Post Request */
+		// Parse Form Data
+		c.Request.ParseForm()
+
+		/* Get Title and secription from the Post request */
+		var title string = form.GetFormValue(c, "title")
+		var description string = form.GetFormValue(c, "description")
+
+		// Get cookie of the id value
+		id, err := c.Cookie("id")
+		if err != nil {
+			panic(err)
+		}
+
+		// Update the entry title and description by id
+		notebook_db.UpdateNoteEntry(db, id, title, description)
+
+		// Redirect to the table view page
+		c.Redirect(http.StatusFound, "/view-notes")
 
 	}
 	return gin.HandlerFunc(fn)
