--- cmd/app/notes/note_actions.go
@@ -1,123 +1,0 @@
-package notes
-
-import (
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-
-	"gorm.io/gorm"
-
-	"notebook_app/cmd/app/notebook_db"
-
-	"notebook_app/cmd/app/user_session"
-
-	"notebook_app/cmd/app/form"
-)
-
-/* Post Requests */
-
-func DeleteOrEditNote(db *gorm.DB) gin.HandlerFunc {
-	fn := func(c *gin.Context) {
-		/* Handle Delete and Edit Button POST Requests
-		on the view notes page */
-		// Parse Form Data
-		c.Request.ParseForm()
-
-		if form.GetFormValue(c, "delete") != "" {
-			/* Delete Note Post request */
-			// get entry id for the deleting an entry
-			id := form.GetFormValue(c, "delete")
-
-			// delete entry
-			//db.Delete(&notebook_db.Note{}, id)
-
-			// set id of entry to edit
-			c.SetCookie("id", id, 10, "/delete-note", c.Request.URL.Hostname(), false, true)
-
-			// redirect to notes view page
-			c.Redirect(http.StatusFound, "/delete-note")
-
-		} else if form.GetFormValue(c, "edit") != "" {
-			/* Edit Note Post Request Redirect */
-			// get entry id for the editing an entry
-			id := form.GetFormValue(c, "edit")
-
-			// set id of entry to edit
-			c.SetCookie("id", id, 10, "/edit-note", c.Request.URL.Hostname(), false, true)
-
-			// redirect to edit note
-			c.Redirect(http.StatusFound, "/edit-note")
-		}
-	}
-	return gin.HandlerFunc(fn)
-}
-
-func AddNewNote(db *gorm.DB) gin.HandlerFunc {
-	fn := func(c *gin.Context) {
-		// Parse Form Data
-		c.Request.ParseForm()
-
-		/* Get Title and description from the Post request */
-		var title string = form.GetFormValue(c, "title")
-		var description string = form.GetFormValue(c, "description")
-
-		// Get User ID Data
-		user_id := user_session.GetUserSessionData(c, "user_id").(uint)
-
-		// Create new note entry
-		notebook_db.CreateNewNoteEntry(db, title, description, user_id)
-
-		// Redirect to the table view page
-		c.Redirect(http.StatusFound, "/view-notes")
-	}
-	return gin.HandlerFunc(fn)
-}
-
-func EditNote(db *gorm.DB) gin.HandlerFunc {
-	fn := func(c *gin.Context) {
-		/* Edit Note Post Request */
-		// Parse Form Data
-		c.Request.ParseForm()
-
-		/* Get Title and secription from the Post request */
-		var title string = form.GetFormValue(c, "title")
-		var description string = form.GetFormValue(c, "description")
-
-		// Get cookie of the id value
-		id, err := c.Cookie("id")
-		if err != nil {
-			panic(err)
-		}
-
-		// Update the entry title and description by id
-		notebook_db.UpdateNoteEntry(db, id, title, description)
-
-		// Redirect to the table view page
-		c.Redirect(http.StatusFound, "/view-notes")
-
-	}
-	return gin.HandlerFunc(fn)
-}
-
-func DeleteNote(db *gorm.DB) gin.HandlerFunc {
-	fn := func(c *gin.Context) {
-		/* Edit Note Post Request */
-		// Parse Form Data
-		c.Request.ParseForm()
-
-		// Get cookie of the id value
-		id, err := c.Cookie("id")
-		if err != nil {
-			panic(err)
-		}
-
-		// delete entry
-		db.Delete(&notebook_db.Note{}, id)
-
-		// Redirect to the table view page
-		c.Redirect(http.StatusFound, "/view-notes")
-
-	}
-	return gin.HandlerFunc(fn)
-}
-
